requirements.txt
streamlit
openai-whisper
torch
numpy
soundfile
transformers
sentencepiece
imageio
imageio-ffmpeg